Make `ReproduciblePropertiesWriter` entry order independent
In order to make the parameter name index files reproducible.

The opportunity was taken to convert `ReproduciblePropertiesWriter` to
Kotlin which required introducing another project to sit below
`:build` as to allow converting the remaining Groovy code in this
project incrementally to Kotlin. This is because Groovy code
cannot depend on Kotlin code from the same project.

When and if all code from `:build` is migrated to Kotlin into
`:buildBase` then `:buildBase` can be renamed to `:build`.